Agregando funcionalidad
All the files in the "public" folder and the "assets" folder were eliminated to a great extent; this in order to make an integration with functionality where everything works through an html file and it is REAC JS who provides the rendering of the information.

Use raccoon-edu-server for http requests.